coords, booking, modify vs update

+++ src/models/activity/activity_update.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 class ActivityUpdate(BaseModel):
     name: Optional[str]
     description: Optional[str]
-    x: Optional[int]
-    y: Optional[int]
+    x: Optional[float]
+    y: Optional[float]
     is_private: Optional[bool]
     is_active: Optional[bool]
